Initial translation of Guyton92 parameters for the Moore94 model.
The parameters and initial state variables for the Moore94 model should be
dependent on those of the Guyton92 model. This initial effort accounts for
most of the relevant values (and ignores the irrelevant ones). However, the
hormonal effects on reabsorption and secretion in the distal nephron are
not yet accounted for.

The default value of the parameter "glmcubic" has also been set to zero.
Linear scaling is clearly the correct approach to use, and this parameter
should be removed at some point.